// src/app/shared-portal/enums/viewLeadsPricing.enums.ts
export enum ProductType {
  NONE = 'Select',
  PAYE = 'PAYE',
  LTD = 'LTD',
  STRATEGY = 'STRATEGY',
}
  
  export enum Dummy {
    Dummy1 = 'Dummy1',
    Dummy2 = 'Dummy2',
    Dummy3 = 'Dummy3',
  }
  
  export enum Fee {
    ThreePercent = '3%',
    FourPercent = '4%',
    FivePercent = '5%',
    TenPercent = '10%',
    FifteenPercent = '15%',
  }